requirements.txt
pandas == 2.0.3
joblib >= 1.0.2
panel == 1.2.3
requests == 2.31.0
azure-storage-blob == 12.16.0
numpy == 1.26.4
scikit-learn == 1.3.0
openpyxl == 3.1.2
